pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ShyamSundhar1411 <clashwithchiefrpjyt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.3"
django-rest-framework = "^0.1.0"
django-restframework = "^0.0.1"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
mypy = "^1.8.0"
pylint = "^3.1.0"
pre-commit = "^3.6.2"
py-xid = "^1.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}
djangorestframework-simplejwt = "^5.3.1"
django-autoslug = "^1.9.9"
django-stubs = "^4.2.7"
djangorestframework-stubs = "^3.14.5"
pylint-django = "^2.5.5"
djangorestframework-camel-case = "^1.4.2"
django-import-export = "^4.3.3"

[tool.flake8]
max-line-length = 88
max-complexity = 12
ignore = ["E402","E501","W503","D100"]

[tool.pylint]
load-plugins = ["pylint_django"]
django-settings-module = "backend.settings"

[tool.pylint.MESSAGE_CONTROL]
disable = """
    too-many-locals,
    line-too-long,
    abstract-method,
    no-member,
    missing-module-docstring,
    missing-class-docstring,
    too-many-ancestors,
    too-few-public-methods,
    missing-function-docstring,
    broad-exception-caught,
    protected-access,
    duplicate-code,
    return-in-init,
    wildcard-import,
    no-name-in-module
"""
enable = "useless-suppression"

[tool.pylint.REPORTS]
reports="no"

[tool.pylint.FORMAT]
max-line-length="88"

[tool.pylint.BASIC]
good-names = "_,i,e,setUp,tearDown,maxDiff"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = ["autoslug", "django", "drf_yasg", "import_export", "phonenumber_field", "rest_framework", "rest_framework_simplejwt", "xid"]

[tool.bandit]
exclude_dirs = ["tests"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[tool.pytest]
DJANGO_SETTINGS_MODULE = "backend.settings"

[tool.mypy]
plugins = ["mypy_django_plugin.main","mypy_drf_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "backend.settings"

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
